Add VietnamRevolts and  broadly expand Modifiers.
Modifiers are not a more integral piece and now gain the ability to
overlay CardPlay with specific functionality. They can be declared by
the player to be used for the scope of a given play in order to receive
an optional benefit (such as +1 op in VietnamRevolts when playing
exclusively in SE Asia).

Changed Moves::CardPlay constructor to accept modifiers.

Potentially deprecated Country#can_add_influence? in favor of consistent
whitelisting.

Added {un}playable? to Modifiers so they can only be played if they do not allow
the game to reach an unplayable state. For instance, the player may
pick a modifier that restricts them to operation solely within SE Asia,
but may not have any method of legally operating there (no influence,
etc).