Add partial implementation of NATO.
Adds the ability to conditionally execute a card's event by providing a
custom implementation of Card#event_playable?. When an event is not
playable, then no Validators or Modifiers are executed. Also adds a hook
for card removal only in cases where the card should be removed per
Card#remove_after_event? and the event has executed.