Add basic underpinnings for a stack-oriented game engine.
Using the stack concept, the game engine can be abstracted away from the
specifics of game-specific concepts such as influence, countries, etc.

The entire game cycle can be expressed in terms of 'Instructions' and
'MoveArbitrators' which themselves can add more items to the stack.

The overall components of the game become:

 - The instruction stack
 - The engine - a class that encapsulates the stack, constantly
   attempting to execute whatever input it is given.
 - The game - a class that contains various sub-components (victory
   track, countries)

State of the game mutates by virtue of instruction execution. Each
instruction is provided context by injection of appropriate variables
from the game layer.

When player input is required (in the form of a 'Move'), a
MoveArbitrator determines the next course of action, and will return the
appropriate instructions to the stack for the engine to execute.