Add card disposal logic & associated instructions.
Transfers card lifecycle management for events onto the events
themselves. Each list of instructions describing an event will now
invoke the correct Discard/Remove/Limbo instruction as appropriate.

Adds an example of Discard/Remove/Limbo usage to a few events.

This addresses issues with events such as Shuttle Diplomacy, where the
card must sit in limbo; and with Star Wars, where a card may either be
removed or discarded depending on details surfaced during event
execution.

Adds tracking for the current card being played, so it may be
discarded or surrendered if events have not already dealt with the card.