Integrate counting and tracking of operations points.
Adds an OpsCounter class to encapsulate the work around counting and
validating operations based on their points. It tracks the base value,
plus any operations modifiers (using class OpsModifier).

Adds observers, a mechanism for observing execution of instructions
inside the engine. This needs to be expanded further for other
components to take advantage.

Adds a Matcher class, the beginning of an implementation for the Match()
component of the game DSL.

Removed "ScoreModifier" in favor of DieRoll & Ops.

Removed Either() from the DSL when used inside a termination clause. An
array of Match() items has the same effect.