Add Guards.
Guards provide a sanity check before instantiating an arbitrator to
ensure the arbitrator can be satisfied through player actions.

Guards are also used for the space race, as no arbitrator is required
when attempting a space race. In this case, they prevent a player from
attempting the space race if they do not meet the entry conditions.