Allow RemoveInfluence arbitrator to terminate early.
When removing influence in certain scenarios, there can be more
influence expected to be removed than actually exists within the
restrictions of the country list.

This commit prevents the arbitrator from getting stuck when there is no
more influence that can be removed, even if the maximum amount allowed
to be removed has not been reached.