Add tooling for determining headline execution order.
Adds a class HeadlineCardSelection that encapsulates the player's card
selection during the headline phase. Instances of this class are
sortable in accordance with the rules in 4.5.C. For the purpose of
sorting using the Comparable interface, lower valued instances are a
higher priority. In other words, calling Array#first on a sorted
collection of these instances will yield the first player/card to be
played.

Adds Card#headline_value and Superpower#headline_priority to assist in
ordering.